// source-code/react-manager-master/src/types/api.ts
export interface Result<T = unknown> {
  code: number
  data: T
  msg: string
}